Moving vimrc from /usr/local/vim to home directory.
On ChromeOS we don't have execute permissions in the home directory so
I originally had the .vimrc in /usr/local/vim. Since that dir was not
part of the git repo we lost track of the changes so I'm temporarily
moving this back to home and adding a comment about where to put vimrc
and how to set up the symlink.